refactor: extract Markdown rendering into internal/markdown package
Move the goldmark configuration and the span-heading renderer out of the
extension package into a dedicated internal/markdown package. The package
exposes New() for the configured goldmark instance and a ToHTML() helper
that both the changelog parser and the store info push now use, removing
the duplicated buffer/convert boilerplate.

No behavior change: store-pushed Markdown still renders headings as
<span class="hN"> instead of <h1>-<h6>.
